BUG-1645 : Move concurrency handling to own class
No functional changes
